Package refactoring [=] BaseApiSteps.java and BaseApiStepsImpl.java [=] BaseAssertionSteps.java [=] .gitignore
