+++ .gitignore
@@ -1,3 +1,3 @@
-.idea/
-target/
+/.idea/
+/target/
 *.iml
