+++ src/main/java/ru/gwp/steps/assertion/BaseAssertionSteps.java
@@ -1,4 +1,4 @@
-package ru.gwp.steps;
+package ru.gwp.steps.assertion;
 
 import io.qameta.allure.Step;
